Fix disabled unit tests in Scap
https://gerrit.wikimedia.org/r/c/mediawiki/tools/scap/+/688351 moved
scap to Python 3.  In that process 3 unit tests were disabled because
Lars did not have time to figure out what was going on with those
tests.  This commit revives the disabled tests.  The underlying issue
is that mock.assert_called() does not exist in Python 3.5 (it arrived
in 3.6).  I replaced calls to assert_called() with an assertion check
on the call_count attribute.  In one case assert_called() was replaced
with assert_called_once_with(), which does exist in Python 3.5.

Bug: T283018
Change-Id: I4227807ae9ebadf9eb89389161bef8b6cd87a738