cdblib.py: Python 3 fixes
Writer.put() had a bug where it was counting the number of characters
in the key/value strings rather than the number of bytes.  This led to
construction of corrupt CDB files during "scap cdb-rebuild".

Test that `cdblib.Writer` produces a valid CDB file by round tripping
the writer results back through the `cdblib.Reader` and comparing to the
original dict object.

Bug: T285345
Change-Id: I0e80cbbe5528988c6a4eae6c5a2c9b7f79e90dd9