Add rsync_cdbs configuration parameter
rsync_cdbs is a new boolean configuration parameter, defaulting to
False.  When False, handling of CDB L10N files is the same as before,
where CDB files are generated on the deploy server, then JSON versions
of those CDBs are created, and all rsync operations exclude the CDB
files but include the JSON files.

When rsync_cdbs is True, CDB files are generated on the deploy servers
and JSON versions are NOT created.  When rsyncing between servers, the
CDB files are included.

Bug: T297326
Change-Id: Ib5dbb0321c24fba28d08128d4de4e29ec74d29c6