config: Ensure environment exists before using it
Scap should not silently fall back to the default configuration if the
user passes an invalid environment as an argument. This check is
implemented in the configuration loading code.

The config_file helper in the test_config.py file is modified a little
to always create the configuration file in a scap/ subdirectory, and
optionally in an environment-specific subdirectory of that as well.
Other configuration tests have always passed in the full path to the
configuration file (so not being in scap/ had no effect), but the full
directory structure is needed when testing file location logic.

Bug: T298274
Change-Id: Ia15fa05997c0f6202c1305c8479bf02fdd4ba41d