Add --abort-git-am-on-fail option to scap apply-patches
Prior to this commit, any failed patch application results in leaving
a repo in the middle of a 'git am' operation which needs to be
manually resolved by the operator.

This commit adds the --abort-git-am-on-fail option to scap
apply-patches to cause it to run "git am --abort" when a patch fails
to apply.  This is primarily intended for use by "scap auto-stage"
(arriving in a future commit).

Change-Id: I7317c960af2c1a774bc8e894556868d6c7346539