scap clean: Add 'auto' mode
If scap clean is passed the string 'auto' instead of a branch name
then inactive train branches created more than 8 days ago will be
automatically selected.

Once this is merged and a new scap release has been deployed, similar
code in the 'stage-train' script (in mediawiki/tools/release) can be
removed and replaced with a call to 'scap clean auto'.

Change-Id: Ib5088cabac101af2d58ea8c96f80caacac725ad9