scap prep auto: Add staging fingerprint support
The staging fingerprint is a list of the top-level repos
(operations/mediawiki-config and mediawiki/core), their branches, and
the SHA1 of the commit they are checked out to. This is the set of
information required to reconstruct the staging directory to a prior
state.

For now scap prep auto will print the fingerprint to stdout at the end
of a successful operation.  It also accepts a --fingerprint argument
to specify the desired checkouts.

Bug: T301417

Change-Id: Iad823baeac0ae4a3f49b8dc44505ebe59d965bc7