scap update-wikiversions: Add --no-check flag
By default, scap update-wikiversions checks to see if the specified
branch has been checked out into the staging directory and will throw
an error if not.  This commit adds a --no-check flag to disable this
check.  This is intended to be used along with "scap prep auto".

Also, don't print a backtrace when checking is enabled (the default)
and the branch has not been checked out yet.

Change-Id: Id315eb9b9ae3d63b901b88cd0f8b26abc5677219