ProgressReporter: Report number of in-flight operations
Add add_in_flight() method to ProgressReporter that callers can use to
advance a counter of the number of in-progress operations.  The usual
calls to add_success() or add_failure() will decrement the counter (if
it has been used).

If add_in_flight() is never called, the output of
ProgressReporter/FancyProgressReporter remains the same as before.

Sample output at the end of an operation:
sync-apaches: 100% (in-flight: 0, ok: 1; fail: 0; left: 0)

Other changes:

 * Removed the _output method of FancyProgressReporter which was just
   a copy of the corresponding method of its super class.

Change-Id: I5ee6a367ed57cc2f7e8ed7ee0ffaa17a92ee3c3f