Don't exclude master hosts from sync-apaches phase
This is a followup to Ia55bbe9e424e820f00284b6bb819a5f401e83eef

If one or more of the masters are included in the list of
mediawiki-installation targets, then run scap pull on those targets
too.  This is necessary since the removal of the sync_common phase.

Change-Id: I8eb0e86d44a759d7c91052f448d98bbe9b697da1