scap clean: Allow re-running even if branch dir doesn't exist
Previously scap clean would immediately terminate if there was no
/srv/mediawiki-staging/php-<vers> dir matching the user-supplied
version.  Now scap clean will keep going even if the directory doesn't
exist locally.  This is to allow it to continue cleaning remote
servers.  This is useful in the case where the sync phase of scap
clean didn't complete successfully in a prior run.

Change-Id: I977960a065ee1c533305616d8352c5e0afc70223