Acquire history log lock in shared mode for read-only access
The previous implementation attempted to acquire the lock in exclusive
mode for read-only access in `history.load()` which is not possible. It
must be acquired in shared mode.

Refactored `utils.open_exclusively()` to use a shared advisory lock when
the mode is 'r' or 'rb', and an exclusive lock for all other modes. The
function is renamed to `utils.open_with_lock()` to reflect the more
general behavior.

Bug: T301417
Change-Id: Id91dd6d2d28631e29ad3ea1906d3921b014c618b