blubber: use `pip` to install Python test dependencies
Prod deps will still be provided by the debian packages. Using `pip`
allows us to use more recent versions of the test deps and to add
deps not available as deb packages.

flake8 version whas pinned to 3.2.1 to match the debian stretch package
version that was being used prior to this commit

Bug: T302905
Change-Id: I25ddffceb7ff00961943630538e2882b86af4a6c