Add AbstractSync._perform_sync()
Introduce AbstractSync._perform_sync() to encapsulate a pattern that
is used several times.  This is in preparation for a subsequent commit
that will report more information about a sync.

_perform_sync() keeps track of which hosts have already been synced to
so that it can avoid re-syncing to them.  This logic was previously
handled manually.

Added utils.list_intersection() to encapsulate some verbose code.
Tests included.

Change-Id: I7f21d9015e9fbb33eda76c989513d2e282621849