pipeline: target Stretch, Buster and Bullseye
The Blubber image is based on Stretch but we have moved most servers to
Buster and surely have some on Bullseye. They each provide different
python 3 versions and different set of dependencies, it might be wise to
run tests against all three flavors.

Adjust the Makefile to let us build and test all the Blubber variants.
The high level targets are:

  make images
  make test
  make clean

One can trigger test for a single distribution eg:

  make test-buster

The images are always rebuild and are using Docker layer caching.

Bug: T308020
Depends-On: I39a0b989d7d7116489e69c1409016cc328b598d4
Change-Id: I0b82007232706ad6e9f54f8a6910668e4443004c