Small improvements to Makefile
Replace echo and false by $(error <message>) which does the same but
with a make builtin. The sole change is the exit code is 2.

Make curl to write the file directly (-o option) instead of shell
redirection.  If curl has an error that prevents an empty file from
being created.

Add options to the `docker run` command which launch tests:
* -t to attach a terminal, makes pytest to output color
* -i allows Ctrl+C to be forwarded from the `docker` cli to the running
  process and thus abort the test run.

Change-Id: I23dc8caa950868cff5351ddbae276245bb36331a