Use `pip` to install Sphinx
Sphinx from Stretch does not support the `autodecorator` directive
whereas the version from Bullseye fails on a duplicate entry when
defining a decorator manually.

* Install Sphinx from Pypi so we benefit from latest improvements.
* Remove sphinxcontrib-actdiag and sphinxcontrib-blockdig upper version
  bound which were put in place for Python 2 compatbility (5986ffb34f)
* Do not install run requirements when building the doc since we get
  them installed in the Blubber image.
* Remove the Sphinx dependencies from Build-Depends.

* fix duplicate entries:

scap.cli has a few function intended to be used as decorators which are
documented with '.. decorator' which causes Sphinx to create a function
documentation. However autodoc already generate entries for each of the
function from docs/api.rst which has `.. automodule:: scap.cli`.

With recent Sphinx versions (such as the one provided by Bullseye), the
duplicate entry raises a warning and fails the build.

Exclude the decorator from autodoc discovery and define them manually.

The generated doc changes from:

  scap.cli.command(*args, **kwargs)[source]

      @scap.cli.command(command_name, help="help text"[, subcommands=False])

To:

  @scap.cli.command(command_name, help="help text"[, subcommands=False])

Change-Id: I88f81d37c7af79c0e53b30089782f1ac0c9cc958