deploy-promote: create wikiversions update patch using user ssh-agent
`scap` relies on a local Keyholder proxy to handle its accessible SSH
keys. At the moment the Keyholder agent doesn't contain a key able to
push changes to Gerrit.

Until such a key is added to Keyholder, we will need to rely on a
user-provided ssh-agent to create the wikiversions update patch in
`operations/mediawiki-config`

Bug: T304557
Change-Id: I9c7bdffbc3034d66bb2947961d7043fa2fe118d6