sync-apaches: Include deploy servers as rsync sources
Before:
During scap sync-world, mediawiki installation targets are told to run
"scap pull" with a list of servers to use as candidate rsync sources.
The /etc/dsh/group/scap-proxies file is used to create that
list.

After:
Now the list of rsync sources is created using
/etc/dsh/group/scap-proxies as well as the list of deploy servers from
/etc/dsh/group/scap-masters.

Impact:
In production this means that two additional servers will be
used to spread the rsync load.

Change notes:
* Renamed _proxy_sync_command() to _base_scap_pull_command() to better
  reflect its purpose.

* Updated the docstrings for _base_scap_pull_command() and
  _apache_sync_command().

Bug: T305466
Change-Id: Ieda49a608fcfdc9700b72a07287f82091b5074a1