Remove the --global option from git lfs calls
On Buster+ systems scap fails to deploy due to the '--global' parameter
added to git lfs, that has been deprecated.
The Machine Learning team is probably the last user of git lfs due
to ORES, that we are trying to deprecate, but for the time being
we'd need to migrate it to Buster and Python 3.7, hence this
code change.

Bug: T303801
Change-Id: Id34e989cb7e3bdbfe57f4e4af998c22f4877e9c2