checks: wait after kill
On a timeout, CheckJob issues a SIGKILL without waiting for the process
to terminate. The subprocess might still be running by the time the Proc
object is deleted. Pytest under Python 3.8+ complains with:

 pytest.PytestUnraisableExceptionWarning:
 Exception ignored in: <function Popen.__del__ at 0x7f19c8c03d30>

 Traceback (most recent call last):
   File "/usr/lib/python3.9/subprocess.py", line 1052, in __del__
     _warn("subprocess %s is still running" % self.pid,
 ResourceWarning: subprocess 21 is still running

The warning comes from Popen.__del__ and later Popen.__exit__ wait for
the process to terminate to avoid zombies.

Change-Id: I991b64c40392b48a8dab185e0c9335d6804a7e8e