scap prep: set ssh push url on all repositories
The repositories are cloned from Gerrit with an https url which is then
used when one wants to push back to Gerrit. We thus have to either:

- set a .netrc file on the deployment server with clear text credential
- configure git to push to the ssh url:

 [url "ssh://gerrit.wikimedia.org:29418"]
	pushInsteadOf = https://gerrit.wikimedia.org/r

When I first paired a deployment with Jaime we have stepped on that mine
and had to add the manual configuration, I am guessing others have a
similar setup.

This change sets the remote push url to use ssh which would let us
remove any .netrc entries people might have and/or stop relying on users
to have a custom git configuration.

Add a test to ensure the git remote configuration is done when cloning
or updating the MediaWiki repositories.

Bug: T306425
Change-Id: I9d059a9d5ec4976defe702b35fcbc4c5e11fe284