install-world: remove concurrency lock
scap process dies with an exception if it tries to acquire an existing
lock file created by a user in a different group, as is the case when
`install-world` is executed by an automated process.

Adding the concurrency lock was meant to serialize the installation
with any `prep` running in parallel, but the chances that the automated
`install-world` and `prep` run at the same time are very small. So it
should be safe to remove the lock here for the time being

Bug: T303559
Change-Id: Ifc1411029b1d0dd84d85b3f3fb850abb347511a0