scap prep: make Gerrit source URL configurable
The scap prep plugin comes with the hardcoded:

  SOURCE_URL=https://gerrit.wikimedia.org/r/

Then we already have a configuration option "gerrit_url" with the same
value.

Remove the hardcoded value and switch to the config switch. That might
benefit train-dev if it wants to "scap prep" from a local Gerrit instead
of the canonical one.

Change-Id: I0c5b41cfa194e9641769b464f3f0d666f51f49ff