get_keyholder_key: Return private key path instead of public key
The ssh IdentityFile option is supposed to point to a private key
file, not a public key file.  The keyholder proxy used in production
tolerates being pointed to a public key but a plain ssh-agent (such as
the one used in train-dev) does not.  Passing a public key file to a
regular ssh-agent results in an error message like the following:

Load key "/etc/keyholder.d/trainbranchbot.pub": invalid format

Change-Id: Ic60e5c7228725e7c1dc9b987c75b4e97dfb25d20