Add auto mode to stage-train
If 'auto' is supplied as the branch name, use 'git ls-remote' on the
mediawiki/core repo to find the latest wmf branch.

If --yes is supplied, don't check for screen or tmux.

Pass --abort-git-am-on-fail to scap apply-patches so that it puts
things in a decent state if patch application fails.  This puts it in
alignment with what "scap prep auto" does.

Bug: T310395
Change-Id: Ibb68ac6ad2250fa65042700c1f11427e53ff6016