startup: warn user if not running from a Python virtual environment
Once we stop installing it using a Debian package, scap will be
expected to be run from a venv. If that is not the case, we let the
user know it's their responsibility the required deps are installed
on the system.

To avoid spamming the user in certain scenarios (i.e. train-dev), the
logic can be overridden by creating a file at
`/etc/scap.skip.venv.check`

Bug: T303559, T310858
Change-Id: I144eab38bc499eaee45c5e064122859f7136f607