serial lock: move default location to `/var/lock`
Scap's `prep` operation used to be able to bootstrap the creation of
the `/srv/mediawiki-staging` dir if necessary. This behavior stopped
working after the default location for the serialization lock was
moved under that dir, as the lock logic fails if the dir is missing.

The new location fixes the problem. It's also a standard path for
process locks and the location where the default, non-serializing
Scap lock is placed

Change-Id: I70b2cdf243d43a8ff81a65f25a74cd1c6976bdb7