serial lock: make sure dir for the lock exists
On startup, make sure the directory where the serializing lock is
placed exists.

This change relies on `stage_dir` ("/srv/mediawiki-staging")
belonging to the deployers group ("deployment") and its setgid flag
being set. This should be the case during normal operation.

When the lock dir is created, it inherits both the group and setgid,
which ensures different deployer users can safely access the lock

Change-Id: I4460ed6c9b3eac18f21ee359b23a82219c8c2413