Sphinx program output must use a relative python
Our documentation includes the `--help` output of scap commands which
require our dependencies for execution. Since `bin/scap` uses the system
python (`/usr/bin/python3`) the dependencies might not be installed
which is the case in a fresh environment.  Invoke the scap binary with
python3 which would work when in the doc virtual environment.

Bug: T315541
Change-Id: Ic30faf3873557143d97ab2f368ae67a4c9edd0f9