backport: Use standard reporter to display merge progress
Instead of continuously polling all changes and displaying their
metadata upon each cycle, only poll changes that haven't yet merged and
display the progress using `log.ProgressReporter`.

Added a trailing spinner to `log.ProgressReporter` that cycles each time
the progress is refreshed. Added a `refresh` method to the reporter that
allows the caller to manually update the progress display during polling
cycles so that the spinner animates and lets the user know the process
is still doing work.

Bug: T315901
Change-Id: Iee5fb010be29896cdfcbd734927dd26c6388527a