checks: Define environment variables for current/done/revs dirs
Define additional environment variables for use in check scripts that
include information about the current rev directory, the done rev
directory (until the `finalize` stage this is the previous rev that
would be used during rollback and is not cleaned up), and the directory
containing all revs.

This will be utilized by phabricator's deployment repo to ensure the
right permissions of config files prior to old rev clean up.

Bug: T313259
Change-Id: I2b4d00ebd4e62a147068d09dec3d38774308b080