Pass --wiki=aawiki and --force-version <version> to l10n related maintenance scripts
Ensure that we always use the same wiki for generating l10n files.

Motivation:

Prior to this commit scap would choose the wiki to pass to
l10n-related maintenance scripts by choosing the (alphabetically)
first wiki that matches the version/branch that we want to update.
The problem with this approach is that each wiki can have its own list
of enabled extensions, and the list of enabled extensions determines
if l10n files need to be rebuilt or not.  As the train advances from
testwikis to group0, the "first wiki" for a particular version may
change (as the list of which wikis are on which versions changes).
This sometimes resulted in undesirably l10n file rebuilds.

Now we take advantage of
https://gerrit.wikimedia.org/r/c/operations/mediawiki-config/+/771001
to force the version of mediawiki software being used to generate
l10n.  We use the same wiki (aawiki) every time to ensure that the
extension list is always the same.

Bug: T316714

Change-Id: I81a97635d2506d014314883d88d77cfbd9822bfa