Support checks running before a stage
This allow to run check scripts BEFORE a stage is started instead of
having them run in the previous stage.  For example if one had to stop a
service before the promote stage, the check has to be added to the
config_deploy stage:

 checks:
   stop_services:
     type: command
     stage: config_deploy
     command: bash -c 'systemctl stop apache2 && systemctl stop php-fpm'

That should instead be done before the start of the promote stage. This
introduces `before` and `after` to checks which lets one adjust the
above check to be run before promote:

 checks:
   stop_services:
     type: command
-    stage: config_deploy
+    before: promote
     command: bash -c 'systemctl stop apache2 && systemctl stop php-fpm'

The old `stage` option is kept for backward compatibility.

Change-Id: I45739076ceb71dfa3017092f52e2abb4536f2699