Remove dependency on distutils.util
The python3-distutils package is installed on some hosts but not all.
It is installed on deploy1002, so "scap install-world" doesn't think
it needs to install any additional python packages in the virtualenv
though they might be needed on some target hosts.

Instead of working on that problem directly, I removed the
distutils.util dependency entirely.  To this end I removed the
_parse_bool method of DeploymentsConfig and now rely on the YAML input
to use proper booleans (true and false, that's it) exclusively.

Bug: T317567
Change-Id: Ib2a08e8c797e83e4e7e40ada8ea5273b98de0f20