devel hacks: Change SCAP to REMOTE_SCAP
cli.get_script_path() is sometimes used so scap can run a subcommand
of itself locally.  At these times the function should return the path
to the scap that is currently running.

cli.get_script_path() is also used to prepare a scap command to run on
remote targets.  In this case the [optional] REMOTE_SCAP (formerly
just SCAP) environment variable can be used to override the path to
scap.  This is useful when running scap from a development tree on the
deploy server.  In that scenario the development tree is not available
on targets, so you can set REMOTE_SCAP=/usr/bin/scap to work around
that.

Each call to cli.get_script_path() has been checked and all those
intended for remote execution have been updated.

Change-Id: If069e916bd1a010b7d72b01907f4b168735d817f