Make DeployLocal._valid_chk() static + test coverage
The method does not rely on any instance state, mark it as a static
method.

Add test to cover how we filter out checks based on the stage being run
and the group of targetted servers.

The test cases each are pytest.param with human friendly ids, they show
up when using pytest -v:

 tests/scap/test_deploy.py::test_DeployLocal_valid_chk
 [When on promote stage, accept promote check]
 [When on fetch stage, reject promote check]
 [When on promote stage, accept promote check for canaries]
 [When on promote stage, reject fetch check for canaries]
 [When doing canaries, accept a check without group]
 [When doing canaries, accept a check for canaries]
 [When doing canaries, reject check for different stage]
 [When doing canaries, reject canaries check for different stage]
 [When doing canaries, reject check for different group]

Change-Id: Id22a46c4e18a3fa2e124b995c614359401b5053e