perform-release: Fix old version handling
With the old code, LAST_TAG was not set globally by the time it was
used at the end of the script.  I think this was because
suggest_new_version is being called in a subshell.

I moved the extraction of the old version tag from suggest_new_version
and made the old version a parameter to it.

Change-Id: Icdb69b14e8b89c40875e445b0befdc3864a3fe76