Add binary image support
Use pyinstaller (https://pyinstaller.org/en/stable/) to build a binary
distribution of scap with all dependencies (including Python and any
referenced shared libraries) bundled.

Build and publish a binary distribution image when a protected tag is
created.  At this time all tags are protected in this scap repo.

The contents of the image can be used to deal with T326668.

Bug: T326668