Load section config when host has a trailing dot
I made the mistake of creating a host named `deployment.localhost.` with
a trailing dot. It is a valid DNS name since the trailing dot represent
the root of the DNS hierarchy. When splitting the fqdn by dot, it causes
an extra empty string to be inserted and cause scap.config to try to
load a section 'localhost.' which does not exist.

Strip the trailing dot from the retrieved fqdn just in case. Add a
couple test to ensure we can load the proper section for a given host
regardless of whether it has a trailing dot or not.

Bug: T329144