Attempt to autorecover from interrupted helm deployments
If helm (and therefore helmfile) is interrupted in the middle of a
deployment, it can end up in a state where subsequent deployments fail
due to the prior deployment being in 'pending-upgrade' state.  The way
to recover from this state is to "helm rollback" the release in
question.

This commit adds detection of this state and attempts to recover.