self-installer: overhaul installation mechanism
Instead of creating a local venv on the deployment server and
syncing that out to all targets, we now rely on the per-distro
wheels generated with every release.

The wheels are extracted on the deploy server and synced over to
the targets (each target gets the right wheels for its distro). Then
a new Scap venv is created on the target and the wheels installed in
it

Bug: T329622