Remove sync-pull-masters step
The sync-pull-masters step runs "scap pull" on all deploy servers
other than the current.  This step is a holdover from a time when
deploy servers needed to have a populated /srv/mediawiki directory in
order to properly act as deploy servers.  This hasn't been the case
for a while so remove this unnecessary operation.  Deploy servers that
are marked as mediawiki installation targets will still run
"scap pull" during the sync-apaches phase.