scap: shorten 'synced to testservers' message
The list of available test servers provides no extra information (the
list of hosts is always the same, and includes all of the X-WM-Debug
options). In addition it's very long (and often the IRC message gets
split into two) and makes searching the SAL for an individual host
useless, so remove it.

Change-Id: I34e0d56d973c1161d2cf7e1ad89f09902a889392