backport: limit what Depends-On changes are considered appropriate
Dependencies for a different branch are not normally relevant in the context of Mediawiki repositories.

The only exception being changes to `operations/mediawiki-config`.
When this repo is targeted as a dependency, the only relevant branch
is the production one (i.e. `master`). When this repo is the
dependant, then we fall back to allowing all branches again.

Aslo refactored the validation code while maintaining semantics otherwise


Bug: T345304