deploy: Fix git-lfs support
Fix git-lfs support in `scap deploy` and `scap deploy-local` by
ensuring that `lfs.url` is set based on the original origin of the repo
being deployed.

`git_binary_manager` is no longer a required configuration setting to
enable `git-fat` or `git-lfs`.  If `git_binary_manager` is not set, the
binary manager will be automatically detected by scanning the
`.gitattributes` file (if any) in the repo being deployed.

Bug: T279509, T324664, T180627
Change-Id: I3d0dd46a4cffe6c89c1d9652ad0575587e267bc9