Remove obsolete defaults for git_server
We had the scap configuration `git_server` set to
`deploy1001.eqiad.wmnet` which has been obsolete for quite a while.

Given the deployment is infrastructure dependent, let it be provided by
Puppet via the global scap.cfg. In the case that is not set, raise an
exception to abort before performing the local deployment.

Bug: T197470