MergeMessageFileLists.php: Use a realm-specific extension-list
Our guidelines for "Writing an extension for deployment"[0] have "deploy
to Beta Cluster" as a pre-production step.

The instructions say to add your extension to our release branching
tool, which means the extension will be on-disk in production for two
weeks before deployment to beta.

The only reason to do this oddity, is because you need to add your
extension to `extension-list` in operations/mediawiki-config to generate
l10n in beta.

This is a problem because that file is shared between production and beta.

After this change, beta and production no longer have to share a single
`extension-list` file, an `extension-list-labs` file can be used
instead.

The future process for releasing an extension then, would be to add your
extension to the `extension-list-labs` file, and add it as a submodule
in the `mediawiki/extensions` project. No need to wait two weeks and no
need to deploy to production before beta.

Removing FIXME for get_realm_specific_filename extensionless files
returning the base filename. All existing uses of this function seem to
have extensions, so this should be a noop.

[0]: <https://www.mediawiki.org/wiki/Writing_an_extension_for_deployment>

Bug: T125678