Allow tests to run on a system that has an /etc/scap.cfg file
The test suite was assuming that /etc/scap.cfg does not exist when
tests run.  This resulted in confusing test failures when running on a
system that does have /etc/scap.cfg.  I resolved this by adding a
use_global_config keyword to config.load().  The default value of the
keyword is True.  Several other methods were updated to take the same
keyword arg.  The test suite uses the new keyword to disable reading
of /etc/scap.cfg.

Change-Id: I5902452bebccfe612bd83f36d8552bff60247bd4