.pipeline/blubber.yaml: Revise how test variant is built
Instead of Blubber's built-in Python support, use pip to install tox,
then allow tox to manage installation of dependencies in the normal
way.  This avoids unnecessary installation of Python packages in the
"system" area, and it avoids conflicts between the various
requirements files (for examples, docs requirements can be in conflict
with test requirements).

Change-Id: I8900abb80a553f5f6c28116481952da3a7bd4959