deploy: set umask followup
For `scap deploy`, set the umask in `_load_config()` which runs
earlier that `main()`.  The call to `self.context.setup()` in
`_load_config()` may result in files/directories being created, so the
umask needs to be set before that.

Also: Remove the `git.is_dir()` check from `main()`.  The real check
happens earlier in `_load_config()` as of
f82a6acbb01485f568e30468a866d204db377290.  It's too late to check in
`main()`.

Bug: T364309
Change-Id: I239edc744829f3fd85c513c3c99002c63e0d50de