checks.py: Discard blank checks
Problem:

`deploy.py` expects objects returned by `checks.load()` to have
`before` and `after` attributes, but the `OverrideCheck` class doesn't
have those attributes.  `OverrideCheck` is not a subtype of
`checks.Check`, so bitrot has creeped in.

Instead of creating an OverrideCheck object when a blank check
definition is seen, just drop the definition and move on.  This
achieves the goals of T149668 without bugs.

Bug: T372921
Bug: T149668
Change-Id: I4ea64b1b8e23d51e325df4ab671b8170912a4170