Revise use of deployment history
Instead of recording unsynced or incomplete information in the history
log at the time of `scap prep` and then updating at the end of `scap
sync-world`, just collect and record information at the end of `scap
sync-world`.  `scap prep` and `deploy-promote` now set the
`scap-prep-point` tag on relevant repos to make it easy to find the
most recently prepped commits excluding security patches.

This way, operations like running the train are properly recorded in
deployment history.  This was not the case before (because train
operations don't use `scap prep`).

Overall, I think it is easier to reason about the deployment history
with these changes.

Change-Id: I4bf99a5609aa2c19f53ee2a7054601cfb54ba80e