scap/__init__.py: Remove cruft
Remove unreferenced `all_applications()` function.  `all_applications()`
has been unreferenced since it was first added by Mukunda in 2016 in
deac80e2bb8926679dd990c6fb1ba8ca3ee42692.

Remove the `any()` call which references a bunch of symbols that were
imported earlier in the file.  The comment on this expression
indicates that the purposes was to use the symbols to avoid unused
import warnings, but I don't get any such warning when the expression
is removed.  It looks like flake8 is smart enough to consider the
entries in `__all__` as symbol references.

Change-Id: I99be8b4c4a87189a69783bb4b3ab4954977772f6