Don't load plugins from user's home directory or current directory
To improve the security of scap (and to make it easier to reason about
what code is loaded into scap), remove the capability of loading
plugins from `~/.scap/plugins` or `$(cwd)/scap/plugins`.  This is
particularly important if running scap under sudo, which we may want
to do to support T376023.

Change-Id: I69a3113b289ebf9c2aa1d7cac0180539d8b91906